requirements.txt
einops==0.6.1
hydra-core==1.3.2
numpy==1.23.5
numpy==1.24.2
omegaconf==2.3.0
pandas==1.5.3
scikit_learn==1.2.2
scipy==1.10.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
